feat(#92): remove redundant code
