--- src/main/java/com/github/lombrozo/jsmith/antlr/rules/SimplifiedSemanticRule.java
@@ -1,65 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2023-2024 Volodya Lombrozo
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package com.github.lombrozo.jsmith.antlr.rules;
-
-import com.github.lombrozo.jsmith.antlr.Context;
-import com.github.lombrozo.jsmith.antlr.semantic.Semantic;
-import com.github.lombrozo.jsmith.antlr.view.Text;
-
-public final class SimplifiedSemanticRule implements Rule {
-
-    private final Rule origin;
-
-    private final Semantic semantic;
-
-    public SimplifiedSemanticRule(final Rule origin, final Semantic semantic) {
-        this.origin = origin;
-        this.semantic = semantic;
-    }
-
-    @Override
-    public Rule parent() {
-        return this.origin.parent();
-    }
-
-    @Override
-    public Text generate(final Context context) {
-        return this.semantic.alter(this.origin.generate(context));
-    }
-
-    @Override
-    public void append(final Rule rule) {
-        this.origin.append(rule);
-    }
-
-    @Override
-    public String name() {
-        return this.origin.name();
-    }
-
-    @Override
-    public Rule copy() {
-        return new SimplifiedSemanticRule(this.origin.copy(), this.semantic);
-    }
-}
